restore nmc
