+++ chirho/observational/handlers/predictive.py
@@ -1,6 +1,3 @@
-import collections
-import math
-import typing
 from typing import Any, Callable, Generic, Mapping, Optional, TypeVar
 
 import pyro
@@ -8,11 +5,10 @@ import torch
 from typing_extensions import ParamSpec
 
 from chirho.indexed.handlers import IndexPlatesMessenger
-from chirho.indexed.ops import get_index_plates, indices_of
+from chirho.indexed.ops import indices_of
 from chirho.observational.handlers.condition import Observations
 from chirho.observational.internals import (
     bind_leftmost_dim,
-    get_importance_traces,
     site_is_delta,
     unbind_leftmost_dim,
 )
@@ -225,115 +221,3 @@ class PredictiveFunctional(Generic[P, T], torch.nn.Module):
                 and not pyro.poutine.util.site_is_subsample(node)
                 and node["infer"].get("_model_predictive_site", False)
             }
-
-
-class BatchedNMCLogMarginalLikelihood(Generic[P, T], torch.nn.Module):
-    r"""
-    Approximates the log marginal likelihood induced by ``model`` and ``guide``
-    using importance sampling at an arbitrary batch of :math:`N`
-    points :math:`\{x_n\}_{n=1}^N`.
-
-    .. math::
-        \log \left(\frac{1}{M} \sum_{m=1}^M \frac{p(x_n \mid \theta_m) p(\theta_m) )}{q_{\phi}(\theta_m)} \right),
-        \quad \theta_m \sim q_{\phi}(\theta),
-
-    where :math:`q_{\phi}(\theta)` is the guide, and :math:`p(x_n \mid \theta_m) p(\theta_m)`
-    is the model joint density of the data and the latents sampled from the guide.
-
-    :param model: Python callable containing Pyro primitives.
-    :type model: torch.nn.Module
-    :param guide: Python callable containing Pyro primitives.
-        Must only contain continuous latent variables.
-    :type guide: torch.nn.Module
-    :param num_samples: Number of Monte Carlo draws :math:`M`
-        used to approximate marginal distribution, defaults to 1
-    :type num_samples: int, optional
-    """
-
-    model: Callable[P, Any]
-    guide: Optional[Callable[P, Any]]
-    num_samples: int
-
-    def __init__(
-        self,
-        model: torch.nn.Module,
-        guide: Optional[torch.nn.Module] = None,
-        *,
-        num_samples: int = 1,
-        max_plate_nesting: Optional[int] = None,
-        data_plate_name: str = "__particles_data",
-        mc_plate_name: str = "__particles_mc",
-    ):
-        super().__init__()
-        self.model = model
-        self.guide = guide
-        self.num_samples = num_samples
-        self._first_available_dim = (
-            -max_plate_nesting - 1 if max_plate_nesting is not None else None
-        )
-        self._data_plate_name = data_plate_name
-        self._mc_plate_name = mc_plate_name
-
-    def forward(
-        self, data: Point[T], *args: P.args, **kwargs: P.kwargs
-    ) -> torch.Tensor:
-        """
-        Computes the log predictive likelihood of ``data`` given ``model`` and ``guide``.
-
-        :param data: Dictionary of observations.
-        :type data: Point[T]
-        :return: Log marginal likelihood at each datapoint.
-        :rtype: torch.Tensor
-        """
-        get_nmc_traces = get_importance_traces(self.model, self.guide)
-
-        with IndexPlatesMessenger(first_available_dim=self._first_available_dim):
-            with BatchedLatents(self.num_samples, name=self._mc_plate_name):
-                with BatchedObservations(data, name=self._data_plate_name):
-                    model_trace, guide_trace = get_nmc_traces(*args, **kwargs)
-            index_plates = get_index_plates()
-
-        plate_name_to_dim = collections.OrderedDict(
-            (p, index_plates[p])
-            for p in [self._mc_plate_name, self._data_plate_name]
-            if p in index_plates
-        )
-        plate_frames = set(plate_name_to_dim.values())
-
-        log_weights = typing.cast(torch.Tensor, 0.0)
-        for site in model_trace.nodes.values():
-            if site["type"] != "sample":
-                continue
-            site_log_prob = site["log_prob"]
-            for f in site["cond_indep_stack"]:
-                if f.dim is not None and f not in plate_frames:
-                    site_log_prob = site_log_prob.sum(f.dim, keepdim=True)
-            log_weights = log_weights + site_log_prob
-
-        for site in guide_trace.nodes.values():
-            if site["type"] != "sample":
-                continue
-            site_log_prob = site["log_prob"]
-            for f in site["cond_indep_stack"]:
-                if f.dim is not None and f not in plate_frames:
-                    site_log_prob = site_log_prob.sum(f.dim, keepdim=True)
-            log_weights = log_weights - site_log_prob
-
-        # sum out particle dimension and discard
-        if self._mc_plate_name in index_plates:
-            log_weights = torch.logsumexp(
-                log_weights,
-                dim=plate_name_to_dim[self._mc_plate_name].dim,
-                keepdim=True,
-            ) - math.log(self.num_samples)
-            plate_name_to_dim.pop(self._mc_plate_name)
-
-        # move data plate dimension to the left
-        for name in reversed(plate_name_to_dim.keys()):
-            log_weights = bind_leftmost_dim(log_weights, name)
-
-        # pack log_weights by squeezing out rightmost dimensions
-        for _ in range(len(log_weights.shape) - len(plate_name_to_dim)):
-            log_weights = log_weights.squeeze(-1)
-
-        return log_weights
